requirements.txt
flake8==3.6.0
mccabe==0.6.1
pkg-resources==0.0.0
pycodestyle==2.4.0
pyflakes==2.0.0
python-magic==0.4.15
